+++ atlasdb-config/src/main/java/com/palantir/atlasdb/config/TimeLockRuntimeConfig.java
@@ -18,15 +18,34 @@ package com.palantir.atlasdb.config;
 
 import org.immutables.value.Value;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.google.common.base.Preconditions;
+import com.palantir.tokens.auth.BearerToken;
 
 @JsonSerialize(as = ImmutableTimeLockRuntimeConfig.class)
 @JsonDeserialize(as = ImmutableTimeLockRuntimeConfig.class)
 @Value.Immutable
 public abstract class TimeLockRuntimeConfig {
+
+    private static final int MIN_TOKEN_LENGTH = 16;
+
     @Value.Default
     public ServerListConfig serversList() {
         return ImmutableServerListConfig.builder().build();
+    }
+
+    /**
+     * Token added to the Authorization header of the requests to Timelock server
+     * to authorize the request for the specified namespace.
+     */
+    @JsonProperty("auth-token")
+    public abstract BearerToken authToken();
+
+    @Value.Check
+    protected void check() {
+        Preconditions.checkState(authToken() != null || authToken().getToken().length() >= MIN_TOKEN_LENGTH,
+                "The auth token must have at least %s characters.", MIN_TOKEN_LENGTH);
     }
 }