+++ docs/source/release_notes/release-notes.rst
@@ -50,6 +50,12 @@ develop
     *    - Type
          - Change
 
+    *    - |new|
+         - Requests sent by ``atlasdb-client`` to Timelock will have the Authorization header set.
+           Its value is retrieved from configs, from the ``auth-token`` field of ``timelockRuntime``.
+           This enables the possibility of authorizing the requests on the server side.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/3077>`__)
+
     *    - |improved|
          - TimeLock Server now exposes a ``startAtlasDbTransaction`` endpoint which locks an immutable timestamp and then gets a fresh timestamp (in a single round-trip call); new TimeLock clients call this endpoint.
            This saves an estimated one TimeLock round-trip of latency when starting a transaction.